Ugly commit! Contains many hours of work in this single commit. Forgot to keep track.. Some of the changes:
- Pathfinding
	* reworked the algorythm to take walls into account
	* works stable, still uses the whole worldvector
	* the ki only uses it for bringing food back so far

- General changes for testing:
	* only one ant object gets created
	* deactivated / activated some output

- Coding reworked:
	* changed int to size_t for positive counting variables
	* removed some dead code

- Bugfixes:
	* Food can be deleted again
	* Missing piece in sourounding level wall fixed

- New Bugs:
	* Ant texture doesnt rotate right

- A lot of other small stuff I cant remember anymore :p
	-> check diff

+++ src/anthill.cpp
@@ -34,10 +34,9 @@ void AntHill::spawnAnt()
     {
         cout<<"Not enough food"<<endl;
     }
-
 }
 
-void AntHill::setHill(int x, int z)
+void AntHill::setHill(size_t x, size_t z)
 {
     if (x < worldvector.size() && z < worldvector.at(0).size())
     {
