+++ src/objectcreator.h
@@ -7,7 +7,7 @@
 #include "square.h"
 
 extern std::vector < std::vector <Square> > worldvector;
-extern int xworldsize, zworldsize;
+extern size_t xworldsize, zworldsize;
 
 
 class ObjectCreator        //class for all classes, Wall-Blocks and Food-Blocks
